Rename meshtastic -> meshwrapper to be consistent with Python code and differentiate from the protobuf package
