+++ go/meshwrapper/neighbor.go
@@ -1,11 +1,11 @@
-package meshtastic
+package meshwrapper
 
 import (
 	"fmt"
 	"time"
 
 	"buf.build/gen/go/meshtastic/protobufs/protocolbuffers/go/meshtastic"
-	"github.com/timendus/meshbot/meshtastic/helpers"
+	"github.com/timendus/meshbot/meshwrapper/helpers"
 )
 
 type Neighbor struct {
