+++ go/meshwrapper/position.go
@@ -1,4 +1,4 @@
-package meshtastic
+package meshwrapper
 
 import (
 	"math"
