+++ go/meshwrapper/stream_interface.go
@@ -1,4 +1,4 @@
-package meshtastic
+package meshwrapper
 
 import (
 	"errors"
